requirements.txt
google-ml-edu
plotly
tensorflow[and-cuda]
matplotlib
pandas
numpy